3차 스프린트 main에 적용

+++ src/calls/calls.controller.ts
@@ -1,4 +1,15 @@
-import { Body, Controller, Delete, Get, Param, Post, Put, Request, UseGuards } from '@nestjs/common';
+import {
+  Body,
+  Controller,
+  Delete,
+  Get,
+  Param,
+  Post,
+  Put,
+  Request,
+  UnauthorizedException,
+  UseGuards,
+} from '@nestjs/common';
 import { CallsService } from './calls.service';
 import { JwtAuthGuard } from 'src/auth/jwt-auth.guard';
 import { CreateCallRequestDto } from './dtos/create-calls.dto';
@@ -19,26 +30,32 @@ export class CallsController {
 
   @Post()
   async createCall(@Request() req, @Body() createCallRequestDto: CreateCallRequestDto): Promise<Call> {
-    const restaurantId = req.user.restaurantId;
-
-    return await this.callsService.createCall(restaurantId, createCallRequestDto);
+    if (req.user.type === 'login') {
+      const restaurantId = req.user.restaurantId;
+      return await this.callsService.createCall(restaurantId, createCallRequestDto);
+    }
+    throw new UnauthorizedException('로그인이 필요한 기능입니다.');
   }
 
-  @Put(':call_id')
+  @Put(':callId')
   async updateCall(
     @Request() req,
-    @Param('call_id') callId: string,
+    @Param('callId') callId: string,
     @Body() updateCallRequestDto: UpdateCallRequestDto,
   ): Promise<Call> {
-    const restaurantId = req.user.restaurantId;
-
-    return await this.callsService.updateCall(restaurantId, callId, updateCallRequestDto);
+    if (req.user.type === 'login') {
+      const restaurantId = req.user.restaurantId;
+      return await this.callsService.updateCall(restaurantId, callId, updateCallRequestDto);
+    }
+    throw new UnauthorizedException('로그인이 필요한 기능입니다.');
   }
 
-  @Delete(':call_id')
-  async deleteCall(@Request() req, @Param('call_id') callId: string): Promise<void> {
-    const restaurantId = req.user.restaurantId;
-
-    return await this.callsService.deleteCall(restaurantId, callId);
+  @Delete(':callId')
+  async deleteCall(@Request() req, @Param('callId') callId: string): Promise<void> {
+    if (req.user.type === 'login') {
+      const restaurantId = req.user.restaurantId;
+      return await this.callsService.deleteCall(restaurantId, callId);
+    }
+    throw new UnauthorizedException('로그인이 필요한 기능입니다.');
   }
 }
