+++ src/menus/menus.controller.ts
@@ -1,4 +1,16 @@
-import { Body, Controller, Delete, Get, Param, Post, Put, Query, Request, UseGuards } from '@nestjs/common';
+import {
+  Body,
+  Controller,
+  Delete,
+  Get,
+  Param,
+  Post,
+  Put,
+  Query,
+  Request,
+  UnauthorizedException,
+  UseGuards,
+} from '@nestjs/common';
 import { MenusService } from './menus.service';
 import { Menu } from './entities/menus.entity';
 import { JwtAuthGuard } from 'src/auth/jwt-auth.guard';
@@ -13,41 +25,49 @@ export class MenusController {
   constructor(private readonly menusService: MenusService) {}
 
   @Get()
-  async getAllMenus(@Request() req, @Query('category_id') categoryId?: string): Promise<GetAllMenusResponseDto[]> {
+  async getAllMenus(@Request() req, @Query('categoryId') categoryId?: string): Promise<GetAllMenusResponseDto[]> {
     const restaurantId = req.user.restaurantId;
 
     return await this.menusService.getAllMenus(restaurantId, categoryId);
   }
 
-  @Get(':menu_id')
-  async getMenu(@Request() req, @Param('menu_id') menuId: string): Promise<GetMenuResponseDto> {
+  @Get(':menuId')
+  async getMenu(@Request() req, @Param('menuId') menuId: string): Promise<GetMenuResponseDto> {
     const restaurantId = req.user.restaurantId;
 
     return await this.menusService.getMenu(restaurantId, menuId);
   }
 
   @Post()
-  async createMenu(@Request() req, @Body() createMenuRequestDto: CreateMenuRequestDto): Promise<Menu> {
-    const restaurantId = req.user.restaurantId;
-
-    return await this.menusService.createMenu(restaurantId, createMenuRequestDto);
+  async createMenu(@Request() req, @Body() createMenuRequestDto: CreateMenuRequestDto): Promise<{ menuId: string }> {
+    if (req.user.type === 'login') {
+      const restaurantId = req.user.restaurantId;
+      const menuId = await this.menusService.createMenu(restaurantId, createMenuRequestDto);
+      return { menuId };
+    }
+    throw new UnauthorizedException('로그인이 필요한 기능입니다.');
   }
 
-  @Put(':menu_id')
+  @Put(':menuId')
   async updateMenu(
     @Request() req,
-    @Param('menu_id') menuId: string,
+    @Param('menuId') menuId: string,
     @Body() updateMenuRequestDto: UpdateMenuRequestDto,
   ): Promise<Menu> {
-    const restaurantId = req.user.restaurantId;
-
-    return await this.menusService.updateMenu(restaurantId, menuId, updateMenuRequestDto);
+    if (req.user.type === 'login') {
+      const restaurantId = req.user.restaurantId;
+      return await this.menusService.updateMenu(restaurantId, menuId, updateMenuRequestDto);
+    }
+    throw new UnauthorizedException('로그인이 필요한 기능입니다.');
   }
 
-  @Delete(':menu_id')
-  async deleteCategory(@Request() req, @Param('menu_id') menuId: string): Promise<void> {
-    const restaurantId = req.user.restaurantId;
+  @Delete(':menuId')
+  async deleteCategory(@Request() req, @Param('menuId') menuId: string): Promise<void> {
+    if (req.user.type === 'login') {
+      const restaurantId = req.user.restaurantId;
 
-    return await this.menusService.deleteMenu(restaurantId, menuId);
+      return await this.menusService.deleteMenu(restaurantId, menuId);
+    }
+    throw new UnauthorizedException('로그인이 필요한 기능입니다.');
   }
 }
